requirements.txt
annotated-doc==0.0.4
openai==2.32.0
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
brotli==1.2.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cssselect2==0.9.0
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.1
fonttools==4.62.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.3.1
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.35
limits==5.8.0
MarkupSafe==3.0.3
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pillow==12.2.0
pluggy==1.6.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydyf==0.12.1
Pygments==2.20.0
pyphen==0.17.2
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
PyYAML==6.0.3
RapidFuzz==3.14.5
requests==2.33.1
requests-toolbelt==1.0.0
slowapi==0.1.9
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
tinycss2==1.5.1
tinyhtml5==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
watchfiles==1.1.1
weasyprint==68.1
webencodings==0.5.1
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
zopfli==0.4.1
zstandard==0.25.0
